refactor(metadata): extract parseType logic into a separated module
